Rename and move JUnit generation files
